cleanup Leaderboard
